intelligent_album/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
mysql-connector-python==8.0.26
Werkzeug==2.3.7 